Move serialization from raw API to presenter.
Increase the isolation of the JSONAPI logic from the rest of the
system.
